Move config into lyrics namespace.
