+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Configs/Generator/Lyrics/RubyTags/Ja/JaRubyTagGeneratorConfigDialog.cs
@@ -1,0 +1,24 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+#nullable disable
+
+using osu.Framework.Bindables;
+using osu.Game.Rulesets.Karaoke.Configuration;
+using osu.Game.Rulesets.Karaoke.Edit.Generator.Lyrics.RubyTags.Ja;
+
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Configs.Generator.Lyrics.RubyTags.Ja
+{
+    public partial class JaRubyTagGeneratorConfigDialog : RubyTagGeneratorConfigDialog<JaRubyTagGeneratorConfig>
+    {
+        protected override KaraokeRulesetEditGeneratorSetting Config => KaraokeRulesetEditGeneratorSetting.JaRubyTagGeneratorConfig;
+
+        protected override GeneratorConfigSection[] CreateConfigSection(Bindable<JaRubyTagGeneratorConfig> current)
+        {
+            return new GeneratorConfigSection[]
+            {
+                new GenericSection(current)
+            };
+        }
+    }
+}
